Add new LocalizationComponent with original name
